+++ frontend/src/app/modules/page/dashboard/abstraction/dashboard.facade.ts
@@ -89,7 +89,8 @@ export class DashboardFacade {
 
   private setInvestigations(): void {
     this.investigationSubscription?.unsubscribe();
-    this.investigationSubscription = this.investigationsService.getReceivedInvestigations(0, 5).subscribe({
+    // TODO: maybe default sorting keep in getReceivedInvestigations ?
+    this.investigationSubscription = this.investigationsService.getReceivedInvestigations(0, 5, ['createdDate', 'desc']).subscribe({
       next: data => this.dashboardState.setInvestigation({ data }),
       error: (error: Error) => this.dashboardState.setInvestigation({ error }),
     });